Add repeating icons component
