+++ src/pages/orgCulture.module.scss
@@ -92,21 +92,6 @@
   }
 }
 
-/* GHIBLI ICONS */
-.ghibli-icons-layout {
-  display: flex;
-  flex-wrap: nowrap;
-  overflow: hidden; /* Hide anything that goes beyond the container */
-  justify-content: center;
-  width: 100%;
-  margin: 0px;
-}
-
-.ghibli-icons {
-  margin: 10px;
-  width: 50px;
-  height: 50px;
-}
 // video
 .video {
   margin: 40px;
